bpo-37132: Add the imath module.
