+++ Misc/NEWS.d/next/Library/2019-06-02-13-56-16.bpo-37132.axawSH.rst
@@ -1,0 +1,1 @@
+Add the :mod:`imath` module.
